iommu/arm-smmu-qcom: Add QC SMMUv2 VA Size quirk for SDM660
Some IOMMUs are getting set-up for Shared Virtual Address, but:
1. They are secured by the Hypervisor, so any configuration change will
   generate a hyp-fault and crash the system
2. This 39-bits Virtual Address size deviates from the ARM System MMU
   Architecture specification for SMMUv2, hence it is non-standard. In
   this case, the only way to keep the IOMMU as the firmware did
   configure it, is to hardcode a maximum VA size of 39 bits (because of
   point 1).

This gives the need to add implementation details bits for at least some
of the SoCs having this kind of configuration, which are at least
SDM630, SDM636 and SDM660.

These implementation details will be enabled on finding the
qcom,sdm630-smmu-v2 compatible.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@somainline.org>
[ Marijn: Rework for 5.12 compatibility ]
Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>