clk: qcom: smd-rpm: Mark clock enabled in clk_smd_rpm_handoff()
The result of clock handoff is that the clock is voted by APSS and
enabled by RPM.  So it should be marked as enabled.

This, combined with .is_enabled/prepared will ultimately cause RPM
clocks that were enabled by the bootloader to actually be shut down
if unused. We can only afford to do so if we have a functioning ICC
driver. An immediate thought to test for that can be to provide an
ICC path to the RPMCC node, however that would create a couple of
circular dependencies between GCC/DISPCC/.../RPMCC and ICC.

The next best thing to do without breaking older device trees is to
add an opt-in DTS property. Do just that.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>
[Konrad: make conditional, explain the consequences]
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>