arm64: dts: qcom: sdm630: Move interconnects from MDP to MDSS node
In 6874f48bb8b0 ("drm/msm: make mdp5/dpu devices master components") it
is assumed that interconnect nodes live under MDSS instead of MDP, but
this isn't currently the the case for sdm630, causing the device to
crash on boot.  After all, the location of this property hasn't settled
/ been set in stone yet given that none of qcom,mdss and qcom,mdp5
bindings in mdp5.txt document it.

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>