pinctrl: pinconf-generic: Add an overridable way to set bias property
We came to a point where we sometimes we support a few dozen boards
with a given SoC. Sometimes, we have to take into consideration
configurations which deviate rather significatly from the reference
or most common designs. In the context of pinctrl, this often comes
down to wildly different pin configurations. While pins, function and
drive-strength are easily overridable, the (mostly) boolean properties
associated with setting bias, aren't. This wouldn't be much of a
problem if they didn't differ between boards so often, preventing us
from having a "nice" baseline setup without inevitably having to go
with an ugly /delete-property/.

Introduce logic to handle bias-type, a property which sets a single
boolean type of bias on the pin (more than one type of BIAS_ does not
make sense, anyway) to make it easily overridable.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>