dt-bindings: cpufreq: cpufreq-qcom-hw: Add MSM8998
MSM8998 (just like SDM660 and possibly a few others) uses CPUFREQ-HW
"v0", which is the same as CPUFREQ-HW v1, except the Operating State
Manager (OSM) must be programmed manually by the operating system,
whereas on v1 it's done by the firmware. This in turn requires us to
specify a few register regions and a clock input by hand. Document it.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>