soc: qcom: Introduce RPM master stats driver
Introduce a driver to query and expose detailed, per-subsystem (as opposed
to the existing qcom_stats driver which exposes SoC-wide data) about low
power mode states of a given RPM master. That includes the APSS (ARM),
MPSS (modem) and other remote cores, depending on the platform
configuration.

This is a vastly cleaned up and restructured version of the same driver
found in msm-5.4.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>