interconnect: qcom: icc-rpm: Control bus rpmcc from icc
The sole purpose of bus clocks that were previously registered with
rpmcc was to convey the aggregated bandwidth to RPM. There's no good
reason to keep them outside the interconnect framework, as it only
adds to the plentiful complexity.

Add the required code to handle these clocks from within SMD RPM ICC.

RPM-owned bus clocks are no longer considered a thing, but sadly we
have to allow for the existence of HLOS-owned bus clocks, as some
(mostly older) SoCs (ab)use these for bus scaling (e.g. MSM8998 and
&mmcc AHB_CLK_SRC).

This is trivially solved with a single *clk, which is filled and used
iff qp.bus_clk_desc is absent and we have a "bus" clock-names entry
in the DT node.

As a happy accident, the core functionality of the interconnect driver
will now be usable without rpmcc.. Not that you'd want to run your kernel
with that disabled..

This change should(tm) be fully compatible with all sorts of old
Device Trees as far as the interconnect functionality goes.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>